refactor method-ref fixes to be used in headless mode. (#1048)
* refactor method-ref fixes to be used in headless mode.
